pyproject.toml
[project]
name = "anima-tasks"
version = "0.1.0"
description = "Pyinvoke task runner for the anima Rust daemon."
requires-python = ">=3.12"
dependencies = [
    "invoke>=2.2.0",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.uv]
package = false
